Link back to bookshelf for other embedded resources
This is my initial attempt to address issue #129 opened by @adamgreig
to have links back to the bookshelf added to the front page. I am
not sure if this is what you were looking for but I thought I would
give it a try. Let me know what you think.