Have linkchecker skip print.html
https://github.com/rust-lang-nursery/mdBook/issues/789 causes
mdbook to generate print.html content with relative URLS that
are still relative to the individual HTML files and doesn't
correct for the fact that print.html isn't in the same folder.

This commit will verify that the relative URLs are correct in all
of the HTML output except for print.html  I am assuming that
people would like to see the images again in the Discovery book
website as they click through the various chapters, even if the
printable version is still broken.

If the PR with this commit is merged, an issue should be created
to track the fact that this change should be reverted once mdbook
handles relative links correctly in its generaton of print.html