appendix: add hint about ST-LINK config version
My F3 seemed to come with the stlink-v2 and not the stlink-v2-1 config
interface which caused the same "Error: open failed" output.

I suggest to hint to try the older version in that case.

I've also separated Cause and Fix headings because I didn't know how to
phrase it otherwise.