Update udev rules to fit better with modern systems.
Updated the udev rules to use the `uaccess` tag instead of the `MODE` and
`GROUP` modifiers, as that is deprecated.

Renamed the udev rule file to start with 69 instead of 99, so it gets
loaded before user access rules are loaded for `uaccess` permissions,
which is run in 73-seat-late.rules (provided by systemd).

Added text showing how to test for these `uaccess` permissions.