microbit version 0.15.0 crates has been released
A new version of the microbit crates has been released.

Cargo.toml
  *  microbit-v2        0.14.0 -> 0.15.0
     Which updated HAL crates to version 0.18.0
  *  microbit           0.14.0 -> 0.15.0

In connection with this I moved all the examples in the book out of book
text and made them project examples, and used an include statement, like
it's done in `my-solution`, which makes checking the examples faster,
and I also cought a couple of typos I made when I updated the text,
which rust-analyser didn't like.

The examples can be run manually with commands like:
`cargo embed --target thumbv7em-none-eabihf --features v2 --example
it-blinks-1`

Don't know if this messes with any automatic tests yet.