`_y = x` explanation concentrate on `x`.
Moving from plural to singular is a better description of what is
actually occurring. After all, only `x` is considered in the example
of use of gdb.

As mentioned by @BartMassey in #482, it is possible that `y` is
initialized. Given that discussing whether something is really
initialized or not is out of scope of this book, or at least chapter,
this part of the sentence is deleted and not replaced.